Updated triangle

+++ Lab 1/main.cpp
@@ -7,24 +7,24 @@ class Shape
 {
 public:
    // pure virtual function providing interface framework.
-   virtual int getArea() = 0;
-   void setWidth(int w)
+   virtual float getArea() = 0;
+   void setWidth(float w)
    {
       width = w;
    }
-   void setHeight(int h)
+   void setHeight(float h)
    {
       height = h;
    }
 protected:
-   int width;
-   int height;
+   float width;
+   float height;
 };
 
 class Triangle: public Shape
 {
 public:
-   int getArea()
+   float getArea()
    {
       return (width * height)/2;
    }
@@ -32,11 +32,11 @@ public:
 
 int main(void)
 {
-   Triangle  Tri;
+    Triangle  Tri;
     Tri.setWidth(5);
-   Tri.setHeight(7);
-   // Print the area of the object.
-   cout << "Total Triangle area: " << Tri.getArea() << endl;
+    Tri.setHeight(7);
+    // Print the area of the object.
+    cout << "Total Triangle area: " << Tri.getArea() << endl;
 
-   return 0;
+    return 0;
 }
